+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.5
+
+* add expand/collapse operation for CalendarController
+
+## 0.0.4
+
+* remove plugin config
+
 ## 0.0.3
 
 * add header builder